==> notebooks/1_benchmark.ipynb <==
# Prediction benchmark: mean calculation (mean num_orders per center/meal combination)

import pandas as pd

train = pd.read_csv('../data/train/train.csv')
test = pd.read_csv('../data/test.csv')
sample_submission = pd.read_csv('../data/sample_submission.csv')

mean_weekly_orders = train.groupby(['center_id','meal_id'],as_index=False)['num_orders'].mean()
mean_weekly_orders.tail()

test_merged = pd.merge(test,mean_weekly_orders,on=['center_id','meal_id'],how='left')
test_merged.tail()

test_merged['num_orders'].isna().sum()

13 (center, meal) combinations are not available in the train set. I tried two imputation strategies (zero and mean) and found out that the latter gives a better score.

sub_1 = test_merged[['id','num_orders']]
sub_1['num_orders'].fillna(sub_0['num_orders'].mean(),inplace=True)

sub_1

# submission score: 70.29
