==> notebooks/3_poisson_regressor.ipynb <==
# Poisson regressor

import pandas as pd
import numpy as np
from sklearn.model_selection import train_test_split
from sklearn.compose import ColumnTransformer
from sklearn.preprocessing import OneHotEncoder, KBinsDiscretizer
from sklearn.pipeline import Pipeline
from sklearn.metrics import mean_squared_log_error
from sklearn.linear_model import PoissonRegressor

sample_submission = pd.read_csv('../data/sample_submission.csv')

### import merged dataframes (refer to notebook 2 for data wrangling steps)

train_merged = pd.read_csv('../data/train/train_merged.csv')
train_merged.tail()

test_merged = pd.read_csv('../data/test_merged.csv')
test_merged.tail()

train_merged.columns

## train & run poisson regressor

X = train_merged[['center_id', 'meal_id', 'checkout_price', 'base_price',
       'emailer_for_promotion', 'homepage_featured', 
       'center_type', 'category', 'cuisine', 'calendar_week']]

y = train_merged['num_orders']

X_train, X_test, y_train, y_test = train_test_split(X, y)

categories = ['center_id', 'meal_id','emailer_for_promotion', 'homepage_featured',
                                                  'center_type', 'category', 'cuisine', 'calendar_week']

metrics = ['base_price','checkout_price']

trans = ColumnTransformer([
    ('categories', OneHotEncoder(drop='first',sparse=False), categories), 
    ('metrics', KBinsDiscretizer(n_bins=10, encode='onehot-dense', strategy='quantile'), metrics)
     ])

model_pipe_poisson = Pipeline([
    ('feature_engineering', trans), 
    ('model', PoissonRegressor(alpha=0.1,max_iter=10000))
                                ])
model_pipe_poisson.fit(X_train, y_train)

y_train_poiss_pred = model_pipe_poisson.predict(X_train)
y_test_poiss_pred = model_pipe_poisson.predict(X_test)

print('train score:',np.sqrt(mean_squared_log_error(y_train, y_train_poiss_pred)))
print('test score:',np.sqrt(mean_squared_log_error(y_test, y_test_poiss_pred)))

# Submission

X_submission = test_merged[['center_id', 'meal_id', 'checkout_price', 'base_price',
                           'emailer_for_promotion', 'homepage_featured', 
                           'center_type', 'category', 'cuisine', 'calendar_week']]

y_submission_pred = model_pipe_poisson.predict(X_submission)

sub_2 = sample_submission
sub_2['num_orders'] = y_submission_pred
sub_2

## submission score: 67.51

Poisson regressor performed a bit better than my benchmark.
