==> notebooks/4_xgb_recursive_forecasting.ipynb <==
# XGboost regressor with lag-1 feature (recursive forecasting)

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
from sklearn.model_selection import train_test_split
from sklearn.compose import ColumnTransformer
from sklearn.preprocessing import OneHotEncoder, KBinsDiscretizer
from sklearn.pipeline import Pipeline
from sklearn.metrics import mean_squared_log_error
import xgboost as xgb

train_merged = pd.read_csv('../data/train/train_merged.csv')
test_merged = pd.read_csv('../data/test_merged.csv')

## New feature: lag-1
Generally, in forecasting models, lagged values are an important feature. In this case I will define lagged value as the target at the last center/meal occurence. That means that the lag can be higher than a week since the center/meal combinations do not occure regularly over the 145 weeks in the train set.

train_merged['lag_1'] = train_merged.groupby(['center_id','meal_id'])['num_orders'].shift()

train_merged['lag_1'].isna().sum()

We can see that we have about 3597 NaNs in the new lag-1 column which corresponds to the number of center/meal combinations in our train set. The target at the first occurence of each combination does not have a lagged value. Let's impute those NaNs with the mean target (over the train set) of the corresponding center/meal combination.

mean_num_orders_per_combination = train_merged.groupby(['center_id','meal_id'],as_index=False)['num_orders'].mean()
mean_num_orders_per_combination.rename(columns={'num_orders':'mean_num_orders'},inplace=True)
mean_num_orders_per_combination.tail()

train_merged = pd.merge(train_merged,mean_num_orders_per_combination, on=['center_id','meal_id'])
train_merged['lag_1'].fillna(train_merged['mean_num_orders'],inplace=True)

X = train_merged[['center_id', 'meal_id', 'checkout_price', 'base_price',
       'emailer_for_promotion', 'homepage_featured',
       'center_type', 'category', 'cuisine', 'calendar_week','lag_1']]

y = train_merged['num_orders']

X_train, X_test, y_train, y_test = train_test_split(X, y)

categories = ['emailer_for_promotion','homepage_featured',
              'center_type','category', 'cuisine']

trans = ColumnTransformer([
    ('num categories', OneHotEncoder(drop='first',sparse=False), categories),],remainder='passthrough')

model_pipe = Pipeline([
    ('feature_engineering', trans), 
    ('model', xgb.XGBRegressor(max_depth = 10, eta=0.2))   
                    ])
model_pipe.fit(X_train, y_train)

y_train_pred = model_pipe.predict(X_train)
y_test_pred = model_pipe.predict(X_test)

len(y_train_pred[y_train_pred<0]), len(y_test_pred[y_test_pred<0])

We can see that we have some negative prediction that have to be set to zero

y_train_pred[y_train_pred<0] = 0
y_test_pred[y_test_pred<0] = 0

print('train score:',np.sqrt(mean_squared_log_error(y_train, y_train_pred)))
print('test score:',np.sqrt(mean_squared_log_error(y_test, y_test_pred)))

## Challenge: there are no lagged values in the test set (submission set)
Approach: recursive forecasting. Each week:
- fill lagged values
- predict target values
- save predictions as lagged values for next week

### Initialize lagged values for the test set (submission set) as last target value for each center/meal combination in the train set

last_num_orders = train_merged.groupby(['center_id','meal_id'],as_index=False)['num_orders'].last()
last_num_orders.rename(columns={'num_orders':'lag_1'},inplace=True)
last_num_orders.tail()

final_test = pd.DataFrame()
last_num_orders = train_merged.groupby(['center_id','meal_id'],as_index=False)['num_orders'].last()
last_num_orders.rename(columns={'num_orders':'lag_1'},inplace=True)

for week in range(146,156):
    
    test_week = test_merged[test_merged['week']==week]
    test_week = pd.merge(test_week,last_num_orders,on=['center_id','meal_id'],how='left')
    
    X_week = test_week[['center_id', 'meal_id', 'checkout_price', 'base_price',
       'emailer_for_promotion', 'homepage_featured', 
       'center_type', 'category', 'cuisine', 'calendar_week','lag_1']]
    
    y_week = model_pipe.predict(X_week)
    y_week[y_week<0] = 0
    test_week['num_orders'] = y_week
    
    
    last_num_orders = test_week[['center_id','meal_id','num_orders']]
    last_num_orders.rename(columns={'num_orders':'lag_1'},inplace=True)
    
    final_test = pd.concat([final_test,test_week])

## Submission

sub_4 = final_test[['id','num_orders']]
sub_4

## Submission score: 59.40
